apps/economic-compass/requirements.txt
yfinance
pandas
requests
openai
python-dotenv
jinja2
beautifulsoup4
lxml
Pillow